Building a calculator that I can use to calculate accumulated value from dividend stocks I own
Hi Guys, I’m starting to create a function(s) or even an object(s) will eventually do calculations that I can use to calculate accumulated value from dividend stocks I own (so you know the goal).

All I am trying to achieve at this time is a prompt for each variable e.g. stockPrice and I’d like to do it as part of a Object and squeeze in as much functionality within the object… that’s the best way right? Layer in related functions within an object right? 

I believe this is a dataset and functionality I’d like in a single related **Object** 
if I was to code correctly, right? 

Cheers
Roger